print classification process for debugging

+++ src/commons/Classifier.cpp
@@ -648,7 +648,7 @@ querySplits, queryKmerList, targetDiffIdxList, targetInfoList, matchBuffer, cout
                             cout<<"\n";
                             print_binary64(64,AminoAcidPart(currentTargetKmer));
                             cout<<"\n";
-                            cout<< getHammingDistanceSum(currentQuery,currentTargetKmer) << endl;
+                            cout<< (int)getHammingDistanceSum(currentQuery,currentTargetKmer) << endl;
                         }
 //@@@@
 //                        cout<<"Q: ";seqIterator.printKmerInDNAsequence(currentQuery);
